+++ jmetal-core/src/main/java/org/uma/jmetal/encoding/solutiontype/IntSolutionType.java
@@ -67,11 +67,11 @@ public class IntSolutionType extends SolutionType implements IntSolutionTypeTemp
     return solution.getDecisionVariables().length ;
   }
 
-  @Override public double getIntUpperBound(Solution solution, int index) {
-    return solution.getDecisionVariables()[index].getUpperBound() ;
+  @Override public int getIntUpperBound(Solution solution, int index) {
+    return (int)solution.getDecisionVariables()[index].getUpperBound() ;
   }
 
-  @Override public double getIntLowerBound(Solution solution, int index) {
-    return solution.getDecisionVariables()[index].getLowerBound() ;
+  @Override public int getIntLowerBound(Solution solution, int index) {
+    return (int)solution.getDecisionVariables()[index].getLowerBound() ;
   }
 }
