+++ jmetal-core/src/main/java/org/uma/jmetal/encoding/solutiontype/IntSolutionTypeTemplate.java
@@ -6,6 +6,6 @@ public interface IntSolutionTypeTemplate {
   public int getIntValue(Solution solution, int index) ;
   public void setIntValue(Solution solution, int index, int value) ;
   public int getNumberOfIntVariables(Solution solution_);
-  public double getIntUpperBound(Solution solution, int index) ;
-  public double getIntLowerBound(Solution solution, int index) ;
+  public int getIntUpperBound(Solution solution, int index) ;
+  public int getIntLowerBound(Solution solution, int index) ;
 }
